faltan test y orden

+++ src/main/java/com/ias/ecommerce/application/services/DeleteProductService.java
@@ -1,12 +1,17 @@
 package com.ias.ecommerce.application.services;
 
 import com.ias.ecommerce.application.domain.IdentificationProduct;
+import com.ias.ecommerce.application.domain.Product;
+import com.ias.ecommerce.application.errors.InputDataError;
+import com.ias.ecommerce.application.errors.ProductNotFound;
 import com.ias.ecommerce.application.model.product.DeleteProductRequest;
 import com.ias.ecommerce.application.model.product.DeleteProductResponse;
 import com.ias.ecommerce.application.ports.in.DeleteProductUseCase;
 import com.ias.ecommerce.application.ports.out.ProductRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.Optional;
 
 public class DeleteProductService implements DeleteProductUseCase {
 
@@ -20,14 +25,24 @@ public class DeleteProductService implements DeleteProductUseCase {
 
     @Override
     public DeleteProductResponse execute(DeleteProductRequest request) {
-        logger.info(request.toString());
         IdentificationProduct identificationProduct = identificationProduct(request.getIdentificationProduct());
+
+        Optional<Product> productOptional = productRepository.getFindById(identificationProduct);
+        if(!productOptional.isPresent()){
+            throw new ProductNotFound(identificationProduct);
+        }
+
+
+
         productRepository.delete(identificationProduct);
-        logger.info("Eliminado");
         return new DeleteProductResponse();
     }
 
-    private IdentificationProduct identificationProduct(String value){
-        return new IdentificationProduct(value);
+    private IdentificationProduct identificationProduct(String value) {
+        try {
+            return new IdentificationProduct(value);
+        }catch (RuntimeException e){
+            throw new InputDataError(e.getMessage());
+        }
     }
 }
